Improve handling multiple connections and better detect SPI reset.
Projects/tools/LinuxHost/ipclib/npi_lnx_ipc.c

This commit improves the handling of multiple
connections by maintaining a list of active
connections. This list is maintained in addition
to the file descriptor used for select(). It is
primarily useful to send AREQ from RNP to all
active connections, without attempting to send to
non socket file descriptors.

Projects/tools/LinuxHost/ipclib/npi_lnx_spi.c

To better detect SPI reset we use the handshake
signaling. By measuring the time it takes for
SRDY to go high after writing the three 0s
which is the POLL. If it takes more than 100ms
the RNP is not responding to a POLL properly,
and it has most likely reset. In the latter
case the RNP will perform handshake when the
host is performing a POLL. We detect this,
and do not attempt to read out a response
from the RNP. Instead we go back to polling.