Dynamic recovery of missed SRDY interrupts
When NPI Server is compiled to use SRDY interrupts there
is a chance that interrupts are missed. This is mitigated by
capturing the perror and enable timeouts. The timeout is set
dynamically to balance system resource usage and performance.

Also added possibility to configure UART baudrate.

Added new compiler for BeagleBone Black. Link to toolchain is
found next to the compiler argument in the makefile.

Minor change to error management. Instead of a huge switch case
we now use a bit to indicate severity of error, and whether or
not the error requires reset.