Fixed race condition in timer resource for client applications
Some applications may want timed events. For this there's a
module that is started by most applications. This module
runs in a separate thread and maintains 16 timers for each thread
subscribing to it. This is very much like OSAL in RemoTI and
other LPRF stacks

This commit also includes improved debugging of Client side.