Improved UART driver response to fast RNP.
UART on SBL is really fast. So fast, that the driver
would sometimes lose the Synchronous Response. So, logic
is added to not wait for signal in case signal has already
arrived.

Also added hal_dbg_ifc* files to avoid compilation errors. Note,
this module is not fully tested and should ONLY be used as reference