Bootloader handshake has no return status for CC2531.
Can rely on the response bit being set instead. It is the same
across platforms. There is also a chance that if the API is
not supported, but just runs through, that the response is set
to success. This could f.ex. happen if the handshake is sent
to an active RNP.