Update to use realtime clock, and updates to NPI-SPI
Made changes to use realtime clock for duration checks, as
well as time stamped log traces.

NPI-SPI:
1 Add back in check for SRDY assert in poll thread. To avoid
calling npi_poll_data() when we acted on an interrupt that was
not for a POLL.
2 Fix how poll timeout is dynamically changed when interrupt
is defined
3 Unlock mutex after signaling poll thread for npi_srdy_H2L_pol
Change duration that is considered long enough to detect reset
from slow SRDY assert.
Added prints in CloseDevice()
More error checking on received bytes
Update to realtime clock for reset detection in npi_poll_data()
Check for errors on lockRetPoll and lockRetSrdy, which includes
preventing unlocking a mutex that is not locked.

Added synchronization for NPI Client SRSP, as well as check
for number of bytes actually sent/received on socket.

Make first intermediate timeout for GPIO faster than the
following.

Updated how logpath is set in NPI Server (npi_lnx_ipc.c).
Fixed potential segmentation faults in printf() and snprintf()